Move index.ts to a src folder to better support building documentation.
- Add more JSDoc to index.ts.
- Update tsconfig.json to only look at the src folder.
